refactor: tests and structure
module:
- cic-go

packages:
- cic/meta
- cic/net
