+++ meta/meta.go
@@ -1,4 +1,4 @@
-package cic_meta
+package meta
 
 import (
 	"github.com/ethereum/go-ethereum/common"
