+++ net/token_index_test.go
@@ -1,4 +1,4 @@
-package cic_net
+package net
 
 import (
 	"context"
@@ -11,13 +11,13 @@ func TestCicNet_TokenIndex_EntryCount(t *testing.T) {
 	name := "Entry count"
 	wantErr := false
 
+	tokenIndex, err := NewCicNet(conf.rpcProvider, w3.A(conf.tokenIndex))
+
+	if err != nil {
+		t.Fatalf("NewCicNet error = %v", err)
+	}
+
 	t.Run(name, func(t *testing.T) {
-		tokenIndex, err := NewCicNet(conf.rpcProvider, w3.A(conf.tokenIndex))
-
-		if err != nil {
-			t.Fatalf("NewCicNet error = %v", err)
-		}
-
 		got, err := tokenIndex.EntryCount(context.Background())
 
 		if (err != nil) != wantErr {
@@ -25,7 +25,7 @@ func TestCicNet_TokenIndex_EntryCount(t *testing.T) {
 		}
 
 		if got.Cmp(big.NewInt(0)) < 1 {
-			t.Fatalf("EntryCount() = %v, want %v", got, 1)
+			t.Errorf("EntryCount() = %v, want %v", got, 1)
 		}
 	})
 }
@@ -59,16 +59,16 @@ func TestCicNet_TokenIndex_AddressAtIndex(t *testing.T) {
 		},
 	}
 
+	tokenIndex, err := NewCicNet(conf.rpcProvider, w3.A(conf.tokenIndex))
+
+	if err != nil {
+		t.Fatalf("NewCicNet error = %v", err)
+	}
+
 	for _, testcase := range tests {
 		tt := testcase
 
 		t.Run(tt.name, func(t *testing.T) {
-			tokenIndex, err := NewCicNet(conf.rpcProvider, w3.A(conf.tokenIndex))
-
-			if err != nil {
-				t.Fatalf("NewCicNet error = %v", err)
-			}
-
 			got, err := tokenIndex.AddressAtIndex(context.Background(), tt.args.index)
 
 			if (err != nil) != tt.wantErr {
@@ -76,7 +76,7 @@ func TestCicNet_TokenIndex_AddressAtIndex(t *testing.T) {
 			}
 
 			if got != tt.address {
-				t.Fatalf("AddressAtIndex = %v, want %v", got, tt.address)
+				t.Errorf("AddressAtIndex = %v, want %v", got, tt.address)
 			}
 		})
 	}
