+++ net/util.go
@@ -1,4 +1,4 @@
-package cic_net
+package net
 
 import (
 	"context"
@@ -8,7 +8,7 @@ import (
 	"math/big"
 )
 
-func (c *CicNet) latestNonce(ctx context.Context, address common.Address) (uint64, error) {
+func (c *CicNet) LastNonce(ctx context.Context, address common.Address) (uint64, error) {
 	var nonce uint64
 
 	err := c.ethClient.CallCtx(
@@ -25,14 +25,9 @@ func (c *CicNet) latestNonce(ctx context.Context, address common.Address) (uint6
 func (c *CicNet) signAndCall(ctx context.Context, input []byte, txData WriteTx) (common.Hash, error) {
 	var txHash common.Hash
 
-	nonce, err := c.latestNonce(ctx, txData.from)
-	if err != nil {
-		return [32]byte{}, err
-	}
-
 	tx, err := types.SignNewTx(&txData.privateKey, c.kitabuSigner, &types.LegacyTx{
 		To:       &txData.to,
-		Nonce:    nonce,
+		Nonce:    txData.nonce,
 		Data:     input,
 		Gas:      txData.gasLimit,
 		GasPrice: big.NewInt(1),
